\name{tch}
\alias{tch}
\docType{data}
\title{
Raw light intensity for two years (2010 and 2013) of Pinus pinaster
}
\description{
This dataset has 2 years (2010 and 2013) each one with 3 raw profiles (in grayscale).
}
\usage{data("tch")}
\format{
  A \code{list} with 2 elements:
  \describe{
    \item{a \code{y2010}}{a list with 3 radii}
    \item{a \code{y2013}}{a list with 3 radii}
}
}
\details{
  ...
}
\source{
  ...
}
\references{
  ...
}
\examples{
names(tch)
names(tch$y2010)
}
\keyword{datasets}
